+++ src/cli.nim
@@ -219,6 +219,8 @@ Examples:
     metavar: "PROFILE", help: "Set the video profile. For h264: high, main, or baseline"),
   OptDef(names: "-preset, --preset", c: cVid, datum: "preset", metavar: "PRESET",
     help: "Set the video encoder's preset (e.g. ultrafast, medium, slow)"),
+  OptDef(names: "-pix_fmt, --pix-fmt", c: cVid, datum: "pix_fmt", metavar: "FMT",
+    help: "Set the output pixel format (e.g. yuv420p, yuv420p10le)"),
   OptDef(names: "--scale", c: cVid, datum: "scale", metavar: "NUM",
     help: "Scale the output video's resolution by NUM factor"),
   OptDef(names: "--no-seek", c: cVid, kind: Flag, datum: "args.noSeek",